Implement new parsing concept/idiom
This ought to make it easier to express e.g. the difference between a
quoted expression involving a function and one involving a type alone.

In the case of the function we want to separately pretty-print the
return type and the parameters, and potentially template args too.

We currently have a "chunks are separated by whitespace" idea which
breaks down when you have things like "const" and "static" in there