Insert newline+indent overlays instead of replacing text wholesale
Previously this code hid all the existing text and generated a new,
formatted version in its place. For usability and the potential for
additional overlays in the future it seemed better to have granular
overlays, in this case one for each newline+indentation added.

The main change architecturally is that the "printer" closure now
produces a list of formatting instructions, each a cons cell whose car
is a position to insert a newline and whose cdr is the amount to
indent afterwards.

This change also revealed some issues with whitespace handling, as we
sometimes dropped or modified it in the previous approach. Now when
the original text has some whitespace we are replacing with a newline,
we place the newline after the whitespace so it becomes invisible,
though being part of the original text, it can still be found or
searched for.