Fix issues identified by package-lint
The distance -> dist change is not really a problem but "distance"
inside a let-form looks like a call to a function that was eliminated
in old versions of Emacs.

I'm omitting fixes for specifically package-related stuff here. I'm
planning to make this work with MELPA and probably straight.el but
want to do all those changes at once.